move validation helper from abc in sources/ to utils
