+++ src/util.ts
@@ -1,3 +1,5 @@
+import { EventSourceSpec } from './schemas';
+
 export function validateEventName(eventName: string): true {
   if (!eventName) {
     throw new Error('Event name cannot be empty');
@@ -31,4 +33,37 @@ export function validateEventName(eventName: string): true {
   }
 
   return true;
+}
+
+export function validateEventNameUniqueness(spec: EventSourceSpec): void {
+  const eventNameMap = new Map<string, string[]>();
+
+  const processEventType = (
+    sourceType: string,
+    events?: { [key: string]: unknown },
+  ): void => {
+    if (!events) {
+      return;
+    }
+
+    Object.keys(events).forEach((eventName) => {
+      const sources = eventNameMap.get(eventName) || [];
+      sources.push(sourceType);
+      eventNameMap.set(eventName, sources);
+    });
+  };
+
+  processEventType('sqs', spec.sqs);
+
+  const duplicates = Array.from(eventNameMap.entries())
+    .filter(([, sources]) => sources.length > 1)
+    .map(([name, sources]) => `'${name}' in ${sources.join(',')}`);
+
+  if (duplicates.length > 0) {
+    throw new Error(
+      duplicates.length === 1
+        ? `Duplicate event name ${duplicates[0]} found in multiple event source types: ${eventNameMap.get(duplicates[0].slice(1, -1))?.join(', ')}. Event names must be unique across all types.`
+        : `Duplicate event names found across event source types: ${duplicates.join(', ')}. Event names must be unique across all types.`,
+    );
+  }
 }
